requirements.txt
flask 
psycopg[binary]
python-dotenv 
azure-storage-blob